+++ scanpipe/views.py
@@ -902,7 +902,7 @@ class ProjectChartsView(ConditionalLoginRequired, generic.DetailView):
             },
             "dependency": {
                 "queryset": project.discovereddependencies,
-                "fields": ["type", "is_runtime", "is_optional", "is_resolved"],
+                "fields": ["type", "is_runtime", "is_optional", "is_pinned"],
             },
         }
 
@@ -1429,7 +1429,6 @@ class CodebaseResourceListView(
         },
         {
             "field_name": "compliance_alert",
-            "condition": scanpipe_app.policies_enabled,
             "filter_fieldname": "compliance_alert",
             "filter_is_right": True,
         },
@@ -1460,11 +1459,6 @@ class CodebaseResourceListView(
             .order_by("path")
         )
 
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["display_compliance_alert"] = scanpipe_app.policies_enabled
-        return context
-
 
 class DiscoveredPackageListView(
     ConditionalLoginRequired,
@@ -1488,7 +1482,6 @@ class DiscoveredPackageListView(
         },
         {
             "field_name": "compliance_alert",
-            "condition": scanpipe_app.policies_enabled,
             "filter_fieldname": "compliance_alert",
         },
         {
@@ -1520,11 +1513,6 @@ class DiscoveredPackageListView(
             .with_resources_count()
             .order_by_package_url()
         )
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["display_compliance_alert"] = scanpipe_app.policies_enabled
-        return context
 
 
 class DiscoveredDependencyListView(
@@ -1576,8 +1564,8 @@ class DiscoveredDependencyListView(
             "filter_fieldname": "is_optional",
         },
         {
-            "field_name": "is_resolved",
-            "filter_fieldname": "is_resolved",
+            "field_name": "is_pinned",
+            "filter_fieldname": "is_pinned",
         },
         {
             "field_name": "is_direct",
@@ -2096,7 +2084,7 @@ class DiscoveredDependencyDetailsView(
                 "resolved_to_package_uid",
                 "is_runtime",
                 "is_optional",
-                "is_resolved",
+                "is_pinned",
                 "is_direct",
             ],
             "icon_class": "fa-solid fa-info-circle",